requirements.txt
pyTelegramBotAPI
gigachat
python-dotenv

